+++ src/app/events/poap/AddUserForEvent.ts
@@ -9,10 +9,11 @@ import EndPOAP from '../../service/poap/EndPOAP';
 import MongoDbUtils from '../../utils/dbUtils';
 
 export default async (oldState: VoiceState, newState: VoiceState): Promise<any> => {
-	if (oldState.channelId === newState.channelId) {
+	if (oldState.channelId === newState.channelId && (oldState.deaf == newState.deaf)) {
 		// user did not change channels
 		return;
 	}
+	
 	const guild: Guild = (oldState.guild != null) ? oldState.guild : newState.guild;
 	const member: GuildMember = (oldState.guild != null) ? oldState.member : newState.member;
 	
@@ -49,47 +50,52 @@ export const addUserToDb = async (
 		// event change is not related to event parameter
 		return;
 	}
-
-	// Check if user joined channel
-	if (newState.channelId === channel.id) {
-		await updateUserForPOAP(member, db, channel, true).catch(e => LogUtils.logError('failed to capture user joined for poap', e));
-	} else {
-		await updateUserForPOAP(member, db, channel, false).catch(e => LogUtils.logError('failed to capture user left for poap', e));
+	if (newState.deaf) {
+		await updateUserForPOAP(member, db, channel, false, true).catch(e => LogUtils.logError('failed to capture user joined for poap', e));
+		return;
 	}
-
+	const hasJoined: boolean = (newState.channelId === channel.id) || !newState.deaf;
+	await updateUserForPOAP(member, db, channel, hasJoined).catch(e => LogUtils.logError(`failed to capture user change for POAP hasJoined: ${hasJoined}`, e));
 	return;
 };
 
 export const updateUserForPOAP = async (
-	member: GuildMember, db: Db, channel: GuildChannel, hasJoined: boolean,
+	member: GuildMember, db: Db, channel: GuildChannel, hasJoined?: boolean, hasDeafened?: boolean,
 ): Promise<any> => {
 	const poapParticipantsDb: Collection = db.collection(constants.DB_COLLECTION_POAP_PARTICIPANTS);
-	const currentDateStr = (new Date()).toISOString();
-
 	const poapParticipant: POAPParticipant = await poapParticipantsDb.findOne({
 		discordServerId: channel.guild.id,
 		voiceChannelId: channel.id,
 		discordUserId: member.user.id,
 	});
+	
+	if (hasDeafened) {
+		Log.debug(`${member.user.tag} | deafened themselves ${channel.name} in ${channel.guild.name}`);
+		await poapParticipantsDb.deleteOne(poapParticipant);
+		return;
+	}
 
 	if (!hasJoined) {
 		Log.debug(`${member.user.tag} | left ${channel.name} in ${channel.guild.name}`);
-		return poapParticipantsDb.updateOne(poapParticipant, {
+		await poapParticipantsDb.updateOne(poapParticipant, {
 			$set: {
 				endTime: (new Date).toISOString(),
 			},
 		});
+		return;
 	}
 
 	if (poapParticipant !== null && poapParticipant.discordUserId === member.user.id) {
 		Log.debug(`${member.user.tag} | rejoined ${channel.name} in ${channel.guild.name}`);
-		return poapParticipantsDb.updateOne(poapParticipant, {
+		await poapParticipantsDb.updateOne(poapParticipant, {
 			$unset: {
 				endTime: null,
 			},
 		});
+		return;
 	}
-
+	
+	const currentDateStr = (new Date()).toISOString();
 	const result: InsertOneWriteOpResult<POAPParticipant> = await poapParticipantsDb.insertOne({
 		discordUserId: member.user.id,
 		discordUserTag: member.user.tag,