+++ CHANGELOG.md
@@ -7,6 +7,7 @@
    - refactor mongodbutils file
    - add POAP schedule error messaging
    - setup POAP claiming for failed delivery
+   - remove deaf users for poap events
 
 ## 1.8.4-RELEASE (2021-11-03)